modify for checkpoint 1
